'#6 Moving flags to calculation bases
