+++ .idea/csv-plugin.xml
@@ -80,6 +80,13 @@
             </Attribute>
           </value>
         </entry>
+        <entry key="\flags\formulas used.csv">
+          <value>
+            <Attribute>
+              <option name="separator" value=";" />
+            </Attribute>
+          </value>
+        </entry>
       </map>
     </option>
   </component>
