+++ calculationengine/tariff.py
@@ -1,6 +1,6 @@
-from flags.cflags.pv_annuity import PresentValues
+from calculationbases.flags.cflags.pv_annuity import PresentValues
 from input.contract import ContractDTO
-from flags.mf_annuity_flags import Flags
+from calculationbases.flags.mf_annuity_flags import Flags
 
 
 class Tariff(ContractDTO):
